DDT-828: Remove CRUD centric features from base services/controllers
