+++ src/main/java/org/dcsa/core/service/impl/ExtendedBaseServiceImpl.java
@@ -7,16 +7,10 @@ import reactor.core.publisher.Flux;
 
 public abstract class ExtendedBaseServiceImpl<R extends ExtendedRepository<T, I>, T, I> extends BaseRepositoryBackedServiceImpl<R, T, I> implements ExtendedBaseService<T, I> {
 
-    public I getIdOfEntity(T entity) {
-        return getRepository().getIdOfEntity(entity);
-    }
-
     public Flux<T> findAllExtended(ExtendedRequest<T> extendedRequest) {
-        return getRepository().countAllExtended(extendedRequest).map(
-                count -> {
-                    extendedRequest.setQueryCount(count); return count;
-                }).thenMany(
-                        getRepository().findAllExtended(extendedRequest)
+        return getRepository().countAllExtended(extendedRequest)
+                .doOnNext(extendedRequest::setQueryCount)
+                .thenMany(getRepository().findAllExtended(extendedRequest)
         );
     }
 }
